requirements.txt
django
psycopg2-binary
djangorestframework
djangorestframework-simplejwt
drf-yasg